Remove entrypoint and data handling.
Shell can't handle envvars with dashes and will strip them out.

This interferes with electricsheep hd itself, so we can't have an entrypoint in shell.

No entrypoint means we can't ensure that the symlink targets exist.

So we have to get rid of the symlinks.

No symlinks mean we can't have /data as a volume.